refactor: extract Outputcard
